Added a new import to the backend (re).
Added basic password sanitation to the create user route (/formSubmission)

Frontend:
Turned the entire login page layout to 2 sections instead of separating them by url endpoint

Using "sanitationError" on the backend, I created a way for the error messages to be displayed in the layout.

Added checkbox with the function to reveal and hide the password field via useState.