Polishing (added comments, extra spacing for readability)
Fixed a bug where the endpoints werent loading.

Fixed a bug where the middleware was running on undesired endpoints.

middleware now runs on endpoints that require logged in user data. This is needed for the middleware to authenticate the logged in user. If the middleware cant authenticate the user for any reason, then the user is redirected to log in again.