Changed base key:  changed to /static/dist
This change ensures all build files and assets are placed in the dist folder on the server, so that the html can reference the assets at /static/dist/...

Added "outDir" key: When "npm run build" is ran, this key makes it so that automatically the build files are placed in this directory. (../../server/static/dist)

Added "emptyOutDir": This just clears out the old dist folder when you run "npm run build". This helps so you dont have outdated or orphaned assets.