Added a check to diffrentiate post or poll when liking a post., When a user likes a post or poll. The 'addlike' route now checks to see if the post that is receiving the like is a post or poll.
The front end accomodates this with conditional render in the LikeAndComment component. This renders a red heart based on if the user liked a post or a poll.

Added the ability to comment on polls

All of this was done on the main feed as well as the profileFeed component.